# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CDEX Internal CRM API",
    "version": "1.0.0",
    "description": "Session-authenticated API for the Caballes-Go Securities, Inc. internal CRM. Lead conversion is transactional: a lead remains as acquisition history and is linked to one canonical contact/client record.",
    "contact": { "name": "CDEX Engineering" }
  },
  "servers": [{ "url": "/api/v1", "description": "API v1 (same origin)" }],
...
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/leads": {
      "get": {
        "tags": ["Leads"], "summary": "List leads", "parameters": [{ "$ref": "#/components/parameters/ListQuery" }],
        "responses": { "200": { "description": "Paginated leads", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListResponse" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["Leads"], "summary": "Create a lead", "requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
        "responses": { "201": { "description": "Lead created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
    "/contacts": {
      "get": { "tags": ["Contacts"], "summary": "List canonical contacts/clients", "parameters": [{ "$ref": "#/components/parameters/ListQuery" }], "responses": { "200": { "description": "Paginated contacts", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListResponse" } } } } } },
      "post": { "tags": ["Contacts"], "summary": "Create a contact/client", "requestBody": { "$ref": "#/components/requestBodies/JsonBody" }, "responses": { "201": { "description": "Contact created" } } }
    },
...
    "/users/{id}/avatar": {
      "get": { "tags": ["Users"], "summary": "Get a user avatar", "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "$ref": "#/components/responses/Image" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "ccrms.sid", "description": "HttpOnly session cookie created after Microsoft Entra ID authentication." }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "ListQuery": { "name": "query", "in": "query", "description": "Backend list query string. Supports search, page, limit, sort, date ranges, and resource-specific filters.", "schema": { "type": "string", "example": "search=alice&page=1&limit=25&status=Active" } }
    },
    "requestBodies": {
      "JsonBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JsonObject" } } } },
...
    },
    "schemas": {
      "JsonObject": { "type": "object", "additionalProperties": true, "description": "Request fields are validated by the endpoint schema layer." },
      "ApiResponse": { "type": "object", "properties": { "data": {}, "success": { "type": "boolean" }, "message": { "type": "string" } } },
      "ListResponse": { "type": "object", "properties": { "data": { "type": "array", "items": { "type": "object", "additionalProperties": true } }, "meta": { "$ref": "#/components/schemas/Pagination" } } },
      "Pagination": { "type": "object", "properties": { "page": { "type": "integer", "minimum": 1 }, "limit": { "type": "integer", "minimum": 1 }, "total": { "type": "integer", "minimum": 0 }, "totalPages": { "type": "integer", "minimum": 0 } } },
      "Error": { "type": "object", "required": ["success", "message"], "properties": { "success": { "type": "boolean", "example": false }, "message": { "type": "string" }, "code": { "type": "string" }, "details": {} } },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Authentication is required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Required permission is missing", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Image": { "description": "Private image stream", "content": { "image/jpeg": {}, "image/png": {}, "image/webp": {} } }
    }
  }
}